+++ packages/view-types/statistical-plots/src/index.js
@@ -1,7 +1,9 @@
 export { CellSetExpressionPlotSubscriber } from './CellSetExpressionPlotSubscriber.js';
 export { CellSetSizesPlotSubscriber } from './CellSetSizesPlotSubscriber.js';
 export { ExpressionHistogramSubscriber } from './ExpressionHistogramSubscriber.js';
+export { DotPlotSubscriber } from './DotPlotSubscriber.js';
 export { FeatureBarPlotSubscriber } from './FeatureBarPlotSubscriber.js';
 export { default as CellSetSizesPlot } from './CellSetSizesPlot.js';
 export { default as CellSetExpressionPlot } from './CellSetExpressionPlot.js';
 export { default as ExpressionHistogram } from './ExpressionHistogram.js';
+export { default as DotPlot } from './DotPlot.js';