+++ packages/vit-s/src/VitS.js
@@ -26,6 +26,7 @@ import {
 } from './view-config-utils.js';
 import { createLoaders } from './vitessce-grid-utils.js';
 import { createGenerateClassName } from './mui-utils.js';
+import { AsyncFunctionsContext } from './contexts.js';
 
 
 /**
@@ -60,6 +61,7 @@ import { createGenerateClassName } from './mui-utils.js';
  * @param {array} props.jointFileTypes Plugin joint file types.
  * @param {array} props.coordinationTypes Plugin coordination types.
  * @param {null|object} props.warning A warning to render within the Vitessce grid,
+ * @param {boolean} props.pageMode Whether to render in page mode. By default, false.
  * provided by the parent.
  */
 export function VitS(props) {
@@ -81,7 +83,10 @@ export function VitS(props) {
     fileTypes: fileTypesProp,
     jointFileTypes: jointFileTypesProp,
     coordinationTypes: coordinationTypesProp,
+    asyncFunctions: asyncFunctionsProp,
     warning,
+    pageMode = false,
+    children,
   } = props;
 
   const viewTypes = useMemo(() => (viewTypesProp || []), [viewTypesProp]);
@@ -93,6 +98,15 @@ export function VitS(props) {
   const coordinationTypes = useMemo(
     () => (coordinationTypesProp || []),
     [coordinationTypesProp],
+  );
+  const asyncFunctions = useMemo(
+    () => Object.fromEntries(
+      // Convert the array of PluginAsyncFunction instances to a mapping
+      // from function type strings to async functions.
+      asyncFunctionsProp?.map(p => ([p.functionType, p.asyncFunction]))
+      || [],
+    ),
+    [asyncFunctionsProp],
   );
 
   const generateClassName = useMemo(() => createGenerateClassName(uid), [uid]);
@@ -219,26 +233,31 @@ export function VitS(props) {
             {...(remountOnUidChange ? ({ key: configKey }) : {})}
           >
             <AuxiliaryProvider createStore={createAuxiliaryStore}>
-              <VitessceGrid
-                success={success}
-                configKey={configKey}
-                viewTypes={viewTypes}
-                fileTypes={fileTypes}
-                coordinationTypes={coordinationTypes}
-                config={configOrWarning}
-                rowHeight={rowHeight}
-                height={height}
-                theme={theme}
-                isBounded={isBounded}
-                stores={stores}
-              />
-              <CallbackPublisher
-                onWarn={onWarn}
-                onConfigChange={onConfigChange}
-                onLoaderChange={onLoaderChange}
-                validateOnConfigChange={validateOnConfigChange}
-                pluginSpecificConfigSchema={pluginSpecificConfigSchema}
-              />
+              <AsyncFunctionsContext.Provider value={asyncFunctions}>
+                <VitessceGrid
+                  pageMode={pageMode}
+                  success={success}
+                  configKey={configKey}
+                  viewTypes={viewTypes}
+                  fileTypes={fileTypes}
+                  coordinationTypes={coordinationTypes}
+                  config={configOrWarning}
+                  rowHeight={rowHeight}
+                  height={height}
+                  theme={theme}
+                  isBounded={isBounded}
+                  stores={stores}
+                >
+                  {children}
+                </VitessceGrid>
+                <CallbackPublisher
+                  onWarn={onWarn}
+                  onConfigChange={onConfigChange}
+                  onLoaderChange={onLoaderChange}
+                  validateOnConfigChange={validateOnConfigChange}
+                  pluginSpecificConfigSchema={pluginSpecificConfigSchema}
+                />
+              </AsyncFunctionsContext.Provider>
             </AuxiliaryProvider>
           </ViewConfigProvider>
         </QueryClientProvider>